+++ acumos/tests/test_metadata.py
@@ -16,4 +16,29 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ===============LICENSE_END=========================================================
-__version__ = '0.6.3'
+"""
+Provides metadata tests
+"""
+import pytest
+
+from acumos.metadata import _gather_requirements, _filter_requirements, Requirements
+
+
+def test_requirements():
+    '''Tests usage of requirement utilities'''
+    req = Requirements(packages=['~/foo', '~/bar/'])
+    assert req.package_names == {'foo', 'bar'}
+
+    reqs = Requirements(reqs=['foo', 'bar', 'baz'], req_map={'foo': 'bing'}, packages=['~/baz'])
+    req_set = _filter_requirements(reqs)
+    assert req_set == {'bing', 'bar'}
+
+    reqs = Requirements(reqs=['PIL', 'scikit-learn', 'keras', 'baz', 'collections', 'time', 'sys'],
+                        req_map={'PIL': 'pillow'}, packages=['~/baz'])
+    req_names, _ = zip(*_gather_requirements(reqs))
+    assert set(map(str.lower, req_names)) == {'pillow', 'scikit-learn', 'keras'}
+
+
+if __name__ == '__main__':
+    '''Test area'''
+    pytest.main([__file__, ])